IMS-VT: Change parameter type of data usage Change Data Usage methods to take parameters of type long instead of int.

IMS-VT: When TTY is ON, do not allow MO VT call
When TTY mode is ON, the user SHOULD NOT be allowed to make
a Mobile Originating (MO) VT call and an UI alert message will
be displayed explaining MO VT call cannot be initiated.

Propagate the call substate changed message to the UI

IMS-VT : Set the flag to start call with speaker on for incoming VT calls
- There seems to be a race condition between when InCallUI
  enters the video mode and sets the speaker phone on and
  when call state changes to active.
- As a result, the audio route gets reset to earpiece as we
  are not setting speaker phone on flag for the incoming
  case. This change fixes that issue.

IMS-VT: Set speaker phone to true only if bluetooth or headset is not connected

IMS-VT: Convert emergency video call to voice call.
Convert emergency video call to voice call.


IMS-VT: Enable Video conferencing.
Enable Video conferencing.


IMS-VT: Check for property before enabling speaker for MT VT calls
Check persist.radio.ims.audio.output before enabling speaker
if property is set to disable speaker then MT VT calls should not
use the speaker.

+++ tests/src/com/android/server/telecom/testapps/TestManagedVideoProvider.java
@@ -52,7 +52,7 @@ public class TestManagedVideoProvider extends Connection.VideoProvider {
         }
 
         @Override
-        public void onCallDataUsageChanged(RemoteConnection.VideoProvider rvp, int dataUsage) {
+        public void onCallDataUsageChanged(RemoteConnection.VideoProvider rvp, long dataUsage) {
             super.onCallDataUsageChanged(rvp, dataUsage);
         }
